requirements.txt
# ============================================
# FILE: requirements.txt
# ============================================
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-mock==3.12.0
pytest-cov==4.1.0
freezegun==1.4.0